Interim Save
(1) Disabled ability to change a player's first name or last name,
(2) Disabled ability to change a team's name
(3) added edit Team functionality. When editing a Team, the only coaching options in the drop down are the existing team head coach or a coach that is not currently assigned to a team.